// src/model/types.ts
export type data = {
    title: string;
    location: string;
    date: string;
    startTime: string;
    endTime: string;
    chiefReferee: string;
    referee: string;
    duration: string;
    notPlaying: string;
};

export type person = { firstName: string; lastName: string; licence: string };